requirements.txt
torch
torchvision
torchaudio
numpy
pandas
matplotlib
scikit-learn
pillow
opencv-python
pyqt5